crimson/os/seastore: support non aligned shadow extent
This commit break the invariant of LBA tree that enables LBA manager
allocates a block size non-aligned extent, which indicates the two
entries in LBA tree could overlap. This overlap allows one laddr map
to multiple paddrs, which means <laddr: 4096, paddr: Dev(0), 4096>
and <laddr: 4097, paddr: Dev(1), 8192> represent the same extent at
laddr 4096, the data located at Dev(0),4096 and Dev(1),8192 should be
same. We can eliminate the GC pressure of the cold device and reduce
the GC pressure of the hot device with this approach.

Signed-off-by: Zhang Song <zhangsong02@qianxin.com>